+++ R/combine_plots.R
@@ -20,10 +20,12 @@
 #' @param ... Currently ignored.
 #' @inheritParams patchwork::wrap_plots
 #'
+#' @autoglobal
+#'
 #' @examples
 #' library(ggplot2)
 #'
-#' # preparing the first plot
+#' # first plot
 #' p1 <- ggplot(
 #'   data = subset(iris, iris$Species == "setosa"),
 #'   aes(x = Sepal.Length, y = Sepal.Width)
@@ -31,7 +33,7 @@
 #'   geom_point() +
 #'   labs(title = "setosa")
 #'
-#' # preparing the second plot
+#' # second plot
 #' p2 <- ggplot(
 #'   data = subset(iris, iris$Species == "versicolor"),
 #'   aes(x = Sepal.Length, y = Sepal.Width)