+++ R/ggbetweenstats.R
@@ -16,13 +16,6 @@
 #'   variable name for `x` will be used.
 #' @param ylab Labels for `y` axis variable. If `NULL` (default),
 #'   variable name for `y` will be used.
-#' @param pairwise.comparisons Logical that decides whether pairwise comparisons
-#'   are to be displayed (default: `TRUE`). Please note that only
-#'   **significant** comparisons will be shown by default. To change this
-#'   behavior, select appropriate option with `pairwise.display` argument. The
-#'   pairwise comparison dataframes are prepared using the
-#'   `pairwise_comparisons` function. For more details
-#'   about pairwise comparisons, see the documentation for that function.
 #' @param p.adjust.method Adjustment method for *p*-values for multiple
 #'   comparisons. Possible methods are: `"holm"` (default), `"hochberg"`,
 #'   `"hommel"`, `"bonferroni"`, `"BH"`, `"BY"`, `"fdr"`, `"none"`.
@@ -34,7 +27,8 @@
 #'
 #'   You can use this argument to make sure that your plot is not uber-cluttered
 #'   when you have multiple groups being compared and scores of pairwise
-#'   comparisons being displayed.
+#'   comparisons being displayed. If set to `"none"`, no pairwise comparisons
+#'   will be displayed.
 #' @param bf.message Logical that decides whether to display Bayes Factor in
 #'   favor of the *null* hypothesis. This argument is relevant only **for
 #'   parametric test** (Default: `TRUE`).
@@ -107,6 +101,8 @@
 #' @seealso \code{\link{grouped_ggbetweenstats}}, \code{\link{ggwithinstats}},
 #'  \code{\link{grouped_ggwithinstats}}
 #'
+#' @autoglobal
+#'
 #' @details For details, see:
 #' <https://indrajeetpatil.github.io/ggstatsplot/articles/web_only/ggbetweenstats.html>
 #'
@@ -149,7 +145,6 @@ ggbetweenstats <- function(data,
                            x,
                            y,
                            type = "parametric",
-                           pairwise.comparisons = TRUE,
                            pairwise.display = "significant",
                            p.adjust.method = "holm",
                            effsize.type = "unbiased",
@@ -192,7 +187,7 @@ ggbetweenstats <- function(data,
                            ...) {
   # data -----------------------------------
 
-  # convert entered stats type to a standard notation
+
   type <- stats_type_switch(type)
 
   # make sure both quoted and unquoted arguments are allowed
@@ -209,7 +204,6 @@ ggbetweenstats <- function(data,
   test <- ifelse(nlevels(data %>% pull({{ x }})) < 3L, "t", "anova")
 
   if (results.subtitle) {
-    # relevant arguments for statistical tests
     .f.args <- list(
       data         = data,
       x            = as_string(x),
@@ -238,7 +232,7 @@ ggbetweenstats <- function(data,
 
   plot <- ggplot(data, mapping = aes({{ x }}, {{ y }})) +
     exec(geom_point, aes(color = {{ x }}), !!!point.args) +
-    exec(geom_boxplot, outlier.shape = NA, !!!boxplot.args) +
+    exec(geom_boxplot, !!!boxplot.args, outlier.shape = NA) +
     exec(geom_violin, !!!violin.args)
 
   # centrality tagging -------------------------------------
@@ -261,7 +255,7 @@ ggbetweenstats <- function(data,
 
   seclabel <- NULL
 
-  if (isTRUE(pairwise.comparisons) && test == "anova") {
+  if (pairwise.display != "none" && test == "anova") {
     mpc_df <- pairwise_comparisons(
       data            = data,
       x               = {{ x }},
@@ -285,7 +279,7 @@ ggbetweenstats <- function(data,
       ggsignif.args    = ggsignif.args
     )
 
-    # preparing the secondary label axis to give pairwise comparisons test details
+    # secondary label axis to give pairwise comparisons test details
     seclabel <- .pairwise_seclabel(
       unique(mpc_df$test),
       ifelse(type == "bayes", "all", pairwise.display)
@@ -325,6 +319,8 @@ ggbetweenstats <- function(data,
 #' @inheritParams .grouped_list
 #' @inheritParams combine_plots
 #' @inheritDotParams ggbetweenstats -title
+#'
+#' @autoglobal
 #'
 #' @seealso \code{\link{ggbetweenstats}}, \code{\link{ggwithinstats}},
 #'  \code{\link{grouped_ggwithinstats}}
@@ -367,10 +363,7 @@ grouped_ggbetweenstats <- function(data,
                                    grouping.var,
                                    plotgrid.args = list(),
                                    annotation.args = list()) {
-  purrr::pmap(
-    .l = .grouped_list(data, {{ grouping.var }}),
-    .f = ggbetweenstats,
-    ...
-  ) %>%
+  .grouped_list(data, {{ grouping.var }}) %>%
+    purrr::pmap(.f = ggbetweenstats, ...) %>%
     combine_plots(plotgrid.args, annotation.args)
 }