+++ R/ggcorrmat.R
@@ -46,6 +46,8 @@
 #'
 #' @inheritSection statsExpressions::corr_test Correlation analyses
 #'
+#' @autoglobal
+#'
 #' @seealso \code{\link{grouped_ggcorrmat}} \code{\link{ggscatterstats}}
 #'   \code{\link{grouped_ggscatterstats}}
 #'
@@ -191,6 +193,8 @@ ggcorrmat <- function(data,
 #' @inheritParams grouped_ggbetweenstats
 #' @inheritDotParams ggcorrmat -title
 #'
+#' @autoglobal
+#'
 #' @seealso \code{\link{ggcorrmat}}, \code{\link{ggscatterstats}},
 #'   \code{\link{grouped_ggscatterstats}}
 #'
@@ -214,10 +218,8 @@ grouped_ggcorrmat <- function(data,
                               grouping.var,
                               plotgrid.args = list(),
                               annotation.args = list()) {
-  purrr::pmap(
-    .l = .grouped_list(data, {{ grouping.var }}),
-    .f = ggcorrmat,
-    ...
-  ) %>% # `guides = "keep"` is needed because each legend will be different
+  .grouped_list(data, {{ grouping.var }}) %>%
+    purrr::pmap(.f = ggcorrmat, ...) %>%
+    # `guides = "keep"` because legends can be different across grouping levels
     combine_plots(guides = "keep", plotgrid.args, annotation.args)
 }