+++ R/ggscatterstats.R
@@ -46,6 +46,8 @@
 #' @seealso \code{\link{grouped_ggscatterstats}}, \code{\link{ggcorrmat}},
 #' \code{\link{grouped_ggcorrmat}}
 #'
+#' @autoglobal
+#'
 #' @details For details, see:
 #' <https://indrajeetpatil.github.io/ggstatsplot/articles/web_only/ggscatterstats.html>
 #'
@@ -108,16 +110,15 @@ ggscatterstats <- function(data,
   # ensure the arguments work quoted or unquoted
   c(x, y) %<-% c(ensym(x), ensym(y))
 
-  # preparing the data frame
+  # data frame
   data %<>% filter(!is.na({{ x }}), !is.na({{ y }}))
 
   # statistical analysis ------------------------------------------
 
   if (results.subtitle) {
-    # convert entered stats type to a standard notation
     type <- stats_type_switch(type)
 
-    # relevant arguments for statistical tests
+
     .f.args <- list(
       data = data,
       x = {{ x }},
@@ -131,7 +132,7 @@ ggscatterstats <- function(data,
     subtitle_df <- .eval_f(corr_test, !!!.f.args, type = type)
     subtitle <- if (!is.null(subtitle_df)) subtitle_df$expression[[1L]]
 
-    # preparing the BF message for null hypothesis support
+    # BF message for null hypothesis support
     if (type == "parametric" && bf.message) {
       caption_df <- .eval_f(corr_test, !!!.f.args, type = "bayes")
       caption <- if (!is.null(caption_df)) caption_df$expression[[1L]]
@@ -143,7 +144,7 @@ ggscatterstats <- function(data,
   # creating jittered positions
   pos <- position_jitter(width = point.width.jitter, height = point.height.jitter)
 
-  # preparing the scatterplot
+  # scatterplot
   plotScatter <- ggplot(data, mapping = aes({{ x }}, {{ y }})) +
     exec(geom_point, position = pos, !!!point.args) +
     exec(geom_smooth, level = conf.level, !!!smooth.line.args, na.rm = TRUE)
@@ -211,6 +212,8 @@ ggscatterstats <- function(data,
 #' @seealso \code{\link{ggscatterstats}}, \code{\link{ggcorrmat}},
 #' \code{\link{grouped_ggcorrmat}}
 #'
+#' @autoglobal
+#'
 #' @inherit ggscatterstats return references
 #' @inherit ggscatterstats return details
 #'
@@ -260,10 +263,7 @@ grouped_ggscatterstats <- function(data,
                                    grouping.var,
                                    plotgrid.args = list(),
                                    annotation.args = list()) {
-  purrr::pmap(
-    .l = .grouped_list(data, {{ grouping.var }}),
-    .f = ggscatterstats,
-    ...
-  ) %>%
+  .grouped_list(data, {{ grouping.var }}) %>%
+    purrr::pmap(.f = ggscatterstats, ...) %>%
     combine_plots(plotgrid.args, annotation.args)
 }